+++ docker.nix
@@ -3,10 +3,10 @@
     openlane-app ? import ./. {}
 }:
 
-with pkgs; let olenv = python3.withPackages(ps:
-    with ps; [ openlane-app ]
-); in dockerTools.streamLayeredImage {
-    name = "donnio/openlane";
+with pkgs; let
+    olenv = python3.withPackages(ps: with ps; [ openlane-app ]);
+in dockerTools.streamLayeredImage {
+    name = "ghcr.io/efabless/openlane2";
     tag = "${openlane-app.version}";
 
     contents = [
@@ -41,19 +41,3 @@ with pkgs; let olenv = python3.withPackages(ps:
         ];
     };
 }
-
-# {
-#     pkgs ? import ./nix/pkgs.nix,
-#     openlane-app ? import ./. {}
-# }:
-
-# with pkgs; let olenv = python3.withPackages(ps:
-#     with ps; [ openlane-app ]
-# ); in dockerTools.buildImage {
-#     name = "efabless/openlane";
-#     tag = "${openlane-app.version}";
-#     copyToRoot = olenv;
-
-#     created = "now";
-#     config.Cmd = [ "/usr/bin/env" "zsh" ];
-# }